requirements.txt
Django==2.2.12
djangorestframework==3.9.0
django-cors-headers==2.4.0
djangorestframework-jwt
psycopg2
xlrd
Pillow
django-filter
sentry-sdk==0.14.1
django_rest_swagger==2.2.0